ppu/asm/ppu_databuf_asm.S: Make process_ppubuf "more accurate"
While working on this loop to fix write descriptors, I noticed that
there's some unusable code here with the second-byte increment of the
databuf pointer.

This can't ever be used becase Y is only one byte wide.  If we
overflowed it in the middle of the loop, we'd be in to trouble, as we'd
wrap back around to the beginning of the databuf and something funky
would happen.

I can't (and probably don't need to) actually fix this shortcoming, but
it should be made clear that the loop is not intended to be able to
process a buffer with more than 256 bytes of data.

At the moment, I think we're capping the buffer quite a bit smaller than
that, so we wouldn't have run into this limitation, but it's good to get
a handle on it while I'm looking at it.