kernel/scall2.c: Put everything in SCALL2_CODE
I moved the execve code into the regular CODE segment when I was
attempting to give user processes 24K to play with, but this isn't
necessary any more, since I'm not paging out the kernel floating page
when we transition to userspace any more.

This gives us a lot more space in CODE, which we need in order to finish
porting the remaing stuff.