kernel/interrupts: Don't store jpmvec on the stack
This is realistically fine for the interrupt use-case but could be
problematic for switching from/to user context since we don't know that
we'll always return back into the context where we started, thereby
pulling the jmpvec back off the stack.

To eliminated concerns about this, just store the jmpvec right above the
rest of the saved zeropage stuff.