lib: Replace bzero with an asm version
This is broken and makes a mess somehow.  Haven't figured out how yet.
Will continue to work on it later.

/* Original commit message below */

I'm not super proud of the reason I'm doing this, but it needs to be
done anyway, so I'll just live with it.

I'm actually committing this because I was seeing totally weird behavior
while dispatching system calls, where things seemed to go off into the
weeds and hit a random BRK instruction sometime after the call returns.

After slaving over the reverse-step instruction for a good part of a
night, I found that what was happening was the trampoline code at 0x300
was getting overwrittne dunring a bzero() call inside of kprintf().  In
fact, I believe it was the first bzero() call, i.e. the big one to zero
the entire printbuf.

Anyway, I stared at this for a while, but it was a pain in the ass to
debug the generated C function.  So, I went ahead and wrote an asm
version, because I knew I would need to at some point anyway, et voila,
it works now.  The open() syscall is dispatched by my toy /init program,
and we end up in the infinite loop in userspace.

Now that I know for sure what the cause here is, I might go back and try
to debug the generated C function, as this seems like a bad bug to hit.
The function is so simple that it's a bit concerning that the compiler
somehow generated something that doesn't work...  It's also possible
that the state just got clobbered some other way, which would also be
good to know.

Either way, I'm committing this for now, because it's a milestone of
sorts.  I might revise it or add to it in the future.